data yano

+++ resources/views/ppmb/struktur/det_lihat.blade.php
@@ -41,12 +41,12 @@
 								<td >{{$struktur->jabatan}}</td>
 
 								<td align="center" class="col-sm-3">
-									<a href="/det_struktur_ppmb/edit/{{$struktur->id_detil}}" onclick="return confirm('Hapus data ?');">
+									<a href="/det_struktur_ppmb/edit/{{$struktur->id_detil}}" >
 										<button type="button" class="btn btn-success btn-square waves-effect waves-square waves-effect" >
 											<i class="material-icons">Edit</i><span class="icon-name"></span>
 										</button>
 									</a>
-									<a href="/det_struktur_ppmb/delete/{{$struktur->id_detil}}">
+									<a href="/det_struktur_ppmb/delete/{{$struktur->id_detil}}" onclick="return confirm('Hapus data ?');">
 										<button type="button" class="btn btn-danger btn-square 
 										waves-effect waves-square waves-effect">
 										<i class="material-icons">Hapus</i><span class="icon-name"></span>
